requirements.txt

flask
requests
pyjwt
